+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - callback in API to dynamically capture a frame using `context.Capture()`.
 - button to delete selected frame.
+- support to print data type `curveData:Float`.
 
 ------------------------